renaming Timer component files
