--- src/components/timer/Timer.spec.js
@@ -1,1 +1,0 @@
-test("", () => {})
